Introduce execute*, execute
When a plan consists of unordered or ordered plans, the corresponding
manifold fns get called, and the resulting value from executing
individual steps is coerced to a manifold stream. When a plan consists
of a single step, its execution can return anything that can be coerced
to a manifold stream. So, execute's only job is to do that conversion.

Technically it only has to do that when a plan consists of a single
step, but this implementation is simpler, and ->source on something that
is already a Manifold source is cheap anyway.